requirements.txt
# Web Framework
fastapi
uvicorn

# WhatsApp API
pywa

# Database
sqlalchemy
pg8000
psycopg2-binary

# Environment
python-dotenv

beautifulsoup4

selenium
webdriver_manager
google-cloud-storage
Jinja2

# Web scraping & HTTP
requests

# Utilities
typing-extensions